+++ matplot_flex/__init__.py
@@ -1,5 +1,6 @@
 from .axes_utils import get_primary_axes
-from .config import AxisConfig, GridConfig, LegendConfig
+from .config import AxisConfig, GridConfig, LegendConfig, LegendItem, LegendPosition
+from .decorators import draw_legend
 from .layout import (
     IS_VISIBLE_FRAME,
     GraphModule,
@@ -11,10 +12,12 @@ from .layout import (
 )
 from .renderers import Renderer, SeriesSpec, render_bar, render_line, render_multi, render_scatter
 from .templates import plot_on_module, plot_template
-from .text_utils import date_formatter, draw_rounded_frame, draw_text, format_params, sci_formatter
+from .text_utils import date_formatter, draw_rounded_frame, draw_text, draw_text_on_fig, format_params, sci_formatter
 
 __all__ = [
     "AxisConfig",
+    "LegendPosition",
+    "LegendItem",
     "GridConfig",
     "LegendConfig",
     "get_primary_axes",
@@ -25,7 +28,9 @@ __all__ = [
     "render_bar",
     "render_multi",
     "draw_text",
+    "draw_text_on_fig",
     "draw_rounded_frame",
+    "draw_legend",
     "format_params",
     "sci_formatter",
     "date_formatter",